+++ bocadillo/__init__.py
@@ -1,12 +1,14 @@
 from .applications import API, App
 from .errors import HTTPError
-from .middleware import Middleware
+from .middleware import ASGIMiddleware, Middleware
 from .recipes import Recipe
 from .response import Response
+from .request import ClientDisconnect
 from .request import Request
 from .staticfiles import static
+from .sse import server_event
 from .templates import Templates
 from .views import view
 from .websockets import WebSocket, WebSocketDisconnect
 
-__version__ = "0.12.0"
+__version__ = "0.12.5"