+++ bocadillo/request.py
@@ -1,7 +1,10 @@
 from json import JSONDecodeError
 from typing import Any, AsyncGenerator
 
-from starlette.requests import Request as _Request
+from starlette.requests import Request as _Request, ClientDisconnect as _CD
+
+
+ClientDisconnect = _CD
 
 
 class Request(_Request):