+++ docs/.vuepress/config.js
@@ -167,7 +167,8 @@ module.exports = {
             "static-files",
             "hooks",
             "background-tasks",
-            "middleware"
+            "middleware",
+            "sse"
           ])
         },
         {
@@ -197,7 +198,12 @@ module.exports = {
         {
           title: "How-To",
           collapsable: false,
-          children: listDir("how-to", ["yaml-media", "middleware", "tortoise"])
+          children: listDir("how-to", [
+            "yaml-media",
+            "middleware",
+            "tortoise",
+            "socketio"
+          ])
         }
       ],
       "/discussions/": [