+++ tests/test_asgi_middleware.py
@@ -1,16 +1,57 @@
-from starlette.middleware.gzip import GZipMiddleware
-
-from bocadillo import App
+from bocadillo import App, ASGIMiddleware
 
 
-def test_if_asgi_middleware_is_applied():
-    app = App(enable_gzip=False)
-    app.add_asgi_middleware(GZipMiddleware, minimum_size=0)
+def test_asgi_middleware(app: App, client):
+    params = None
+    received_app = False
+    called = False
+
+    class Middleware(ASGIMiddleware):
+        def __init__(self, inner, app: App, **kwargs):
+            super().__init__(inner, app)
+            nonlocal params, received_app
+            params = kwargs
+            received_app = isinstance(app, App)
+
+        def __call__(self, scope):
+            nonlocal called
+            called = True
+            return super().__call__(scope)
+
+    app.add_asgi_middleware(Middleware, hello="world")
+    assert received_app
+    assert params == {"hello": "world"}
 
     @app.route("/")
     async def index(req, res):
         pass
 
-    response = app.client.get("/", headers={"Accept-Encoding": "gzip"})
-    assert response.status_code == 200
-    assert response.headers["content-encoding"] == "gzip"
+    client.get("/")
+    assert called
+
+
+def test_pure_asgi_middleware(app: App, client):
+    initialized = False
+    called = False
+
+    class Middleware:
+        def __init__(self, inner):
+            nonlocal initialized
+            self.inner = inner
+            initialized = True
+
+        def __call__(self, scope: dict):
+            nonlocal called
+            called = True
+            return self.inner(scope)
+
+    app.add_asgi_middleware(Middleware)
+
+    assert initialized
+
+    @app.route("/")
+    async def index(req, res):
+        pass
+
+    client.get("/")
+    assert called