+++ tests/test_request.py
@@ -7,7 +7,7 @@ from bocadillo.constants import ALL_HTTP_METHODS
 
 
 @pytest.mark.parametrize("method", ALL_HTTP_METHODS)
-def test_method(app: App, method: str):
+def test_method(app: App, client, method: str):
     req_method = None
 
     @app.route("/")
@@ -16,7 +16,7 @@ def test_method(app: App, method: str):
             nonlocal req_method
             req_method = req.method
 
-    r = getattr(app.client, method.lower())("/")
+    r = getattr(client, method.lower())("/")
     assert req_method == method
 
 
@@ -31,7 +31,7 @@ def test_method(app: App, method: str):
         ("is_secure", False),
     ],
 )
-def test_url(app: App, attr, value):
+def test_url(app: App, client, attr, value):
     url = None
 
     @app.route("/foo")
@@ -39,7 +39,7 @@ def test_url(app: App, attr, value):
         nonlocal url
         url = req.url
 
-    app.client.get("/foo", params={"bar": "err"})
+    client.get("/foo", params={"bar": "err"})
     assert url is not None
 
     if attr is None:
@@ -48,7 +48,7 @@ def test_url(app: App, attr, value):
         assert getattr(url, attr) == value
 
 
-def test_headers(app: App):
+def test_headers(app: App, client):
     @app.route("/")
     async def index(req, res):
         # Key access.
@@ -62,11 +62,11 @@ def test_headers(app: App):
         # Defaults.
         assert req.headers.get("x-bar", 1) == 1
 
-    r = app.client.get("/", headers={"X-Foo": "foo"})
+    r = client.get("/", headers={"X-Foo": "foo"})
     assert r.status_code == 200
 
 
-def test_query_params(app: App):
+def test_query_params(app: App, client):
     @app.route("/")
     async def index(req, res):
         assert req.url.query == "a=alpha&b=beta&b=BETA&b=Beta"
@@ -85,25 +85,23 @@ def test_query_params(app: App):
         assert req.query_params["b"] == last_item
         assert req.query_params.getlist("b") == ["beta", "BETA", "Beta"]
 
-    r = app.client.get(
-        "/", params={"a": "alpha", "b": ["beta", "BETA", "Beta"]}
-    )
+    r = client.get("/", params={"a": "alpha", "b": ["beta", "BETA", "Beta"]})
     assert r.status_code == 200
 
 
-def test_raw_body(app: App):
+def test_raw_body(app: App, client):
     @app.route("/")
     async def index(req, res):
         body = await req.body()
         assert isinstance(body, bytes)
         res.content = body
 
-    r = app.client.get("/", data="hello")
+    r = client.get("/", data="hello")
     assert r.status_code == 200
     assert r.content == b"hello"
 
 
-def test_form_body(app: App):
+def test_form_body(app: App, client):
     @app.route("/")
     async def index(req, res):
         form = await req.form()
@@ -111,25 +109,25 @@ def test_form_body(app: App):
         # not actually a dict, so not immediately JSON-serializable
         res.media = dict(form)
 
-    r = app.client.get("/", data={"foo": "bar"})
+    r = client.get("/", data={"foo": "bar"})
     assert r.status_code == 200
     assert r.json() == {"foo": "bar"}
 
 
 @pytest.mark.parametrize("data, status", [("{", 400), ("{}", 200)])
-def test_json_body(app: App, data: str, status: int):
+def test_json_body(app: App, client, data: str, status: int):
     @app.route("/")
     class Index:
         async def post(self, req, res):
             res.media = await req.json()
 
-    assert app.client.post("/", data=data).status_code == status
+    assert client.post("/", data=data).status_code == status
 
 
 @pytest.mark.parametrize(
     "get_stream", [lambda req: req, lambda req: req.stream()]
 )
-def test_stream_request(app: App, get_stream):
+def test_stream_request(app: App, client, get_stream):
     @app.route("/")
     class Index:
         async def get(self, req, res):
@@ -147,5 +145,5 @@ def test_stream_request(app: App, get_stream):
         for _ in range(3):
             yield message
 
-    response = app.client.get("/", data=stream())
+    response = client.get("/", data=stream())
     assert response.json() == [message] * 3